+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemConfigurationTest.java
@@ -72,7 +72,6 @@ public class GoogleHadoopFileSystemConfigurationTest {
           put("fs.gs.status.parallel.enable", true);
           put("fs.gs.lazy.init.enable", false);
           put("fs.gs.block.size", 64 * 1024 * 1024L);
-          put("fs.gs.implicit.dir.infer.enable", true);
           put("fs.gs.glob.algorithm", GlobAlgorithm.CONCURRENT);
           put("fs.gs.max.requests.per.batch", 15L);
           put("fs.gs.batch.threads", 15);